move legacy FPCore context to top-level
