+++ fpy2/frontend/__init__.py
@@ -6,5 +6,4 @@ for the FPy language.
 """
 
 from .fpc import fpcore_to_fpy
-from .fpc_context import FPCoreContext
 from .parser import Parser
